Reorganize cypress test files.
